comm: fix send() failure when data exceeds socket send buffer
This patch fixes the send function to not failed when the data to send
is bigger than the internal send buffer of the socket. Previously,
large sends (e.g. during a 'lib scan' with over 1500 slots) could fail
due to partial writes being ignored, resulting in incomplete transmission
and broken behavior. (https://github.com/phobos-storage/phobos/issues/36)

This is a backport of the patch added in Phobos 3.1.

Change-Id: Idaf151bc0ebeaa4ebab0e298b2e2a832122e7463
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.externe@ayudata.fr>