lrs: add a max write per grouping to fifo scheduler
We add a configurable maximum of writes in parallel using the same
grouping for one fifo scheduler. This allows to limit the number
of media that are used in parallel to the same grouping.

If there is already as many current writes of the same grouping as
the configured maximum, new writes of this grouping coming to the
tail of the fifo queue are pushed back to the head. Other newer
requests may be scheduled before the requeued ones. In addition of
breaking the fifo order, this new behaviour could produce additional
tapes movements because new requests could ask for new media that
will remove current media of the targeted grouping of the requeued
write before it comes back to the tail of the fifo queue.

That's why this new limit parameter has to be used carefully, until
we develop a true "per grouping write scheduler", similar to the
grouped read scheduler.

This patch is a first answer to the github issue #28 before we code
a "per grouping write scheduler".
(https://github.com/phobos-storage/phobos/issues/28) .

This is a backport of the patch added in Phobos 3.1. The backport has
removed the features introduced by the commit (5787103 : lrs:
dev_picker: forbid multiple parallel put to same group) because it can
now be done by using the fifo_max_write_per_grouping parameter.

Change-Id: I19afde697beb9c2ef47804ee3ea7657693f2a1f0
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.externe@ayudata.fr>